% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{exportData}
\alias{exportData}
\title{Export data from ClustVis.}
\usage{
exportData(x)
}
\arguments{
\item{x}{structure returned by the function \code{importData}, \code{processData}, \code{generatePCA} or \code{generateHeatmap}.}
}
\value{
a structure depending on the class of the input object
}
\description{
This function allows to export textual data, similar to but not limited to the options on the 'Export' tab of the online ClustVis.
}
