% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{generateHeatmap}
\alias{generateHeatmap}
\title{Generate ClustVis heatmap.}
\usage{
generateHeatmap(proc, showImputed = TRUE, transpose = FALSE,
  clustDistRows = "correlation", clustMethodRows = "average",
  treeOrderingRows = NA, nbrClustersRows = 1,
  clustDistCols = "correlation", clustMethodCols = "average",
  treeOrderingCols = NA, nbrClustersCols = 1, colorAnnoRow = NA,
  colorAnnoCol = NA, legendColorScheme = "Set1", plotWidth = 25,
  plotRatio = 0.8, colorRangeMin = NA, colorRangeMax = NA,
  matrixColorScheme = "RdBu", revScheme = TRUE,
  cellBorder = "grey60", fontSizeGeneral = 10, showNumbers = FALSE,
  fontSizeNumbers = 12, precisionNumbers = 2, showRownames = TRUE,
  fontSizeRownames = NA, showColnames = TRUE, fontSizeColnames = NA,
  showAnnoTitlesRow = TRUE, showAnnoTitlesCol = TRUE,
  maxAnnoLevels = 50, ...)
}
\arguments{
\item{proc}{structure returned by the \code{processData} function.}

\item{showImputed}{whether to show imputed values (default) or missing values on the heatmap.}

\item{transpose}{whether to transpose the heatmap before plotting or not (default).}

\item{clustDistRows}{clustering distance measure used for rows. One from \code{"correlation"} (default), \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, \code{"minkowski"} or \code{NA} for no clustering.}

\item{clustMethodRows}{linkage criterion used for rows. One from \code{"single"}, \code{"complete"}, \code{"average"}  (default), \code{"mcquitty"}, \code{"median"}, \code{"centroid"}, \code{"ward.D2"} or \code{"ward.D"}.}

\item{treeOrderingRows}{how to order branches of the clustering tree of rows to make it visually more attractive. One from \code{"higher median value first"}, \code{"higher mean value first"}, \code{"lower median value first"}, \code{"lower mean value first"}, \code{"original"}, \code{"reverse original"}, or \code{NA} (default) to leave the ordering unchanged.}

\item{nbrClustersRows}{number of clusters in rows separated by gaps.}

\item{clustDistCols}{clustering distance measure used for columns. One from \code{"correlation"} (default), \code{"correlation"}, \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, \code{"minkowski"} or \code{NA} for no clustering.}

\item{clustMethodCols}{linkage criterion used for columns. One from \code{"single"}, \code{"complete"}, \code{"average"}  (default), \code{"mcquitty"}, \code{"median"}, \code{"centroid"}, \code{"ward.D2"} or \code{"ward.D"}.}

\item{treeOrderingCols}{how to order branches of the clustering tree of columns to make it visually more attractive. One from \code{"higher median value first"}, \code{"higher mean value first"}, \code{"lower median value first"}, \code{"lower mean value first"}, \code{"original"}, \code{"reverse original"}, or \code{NA} (default) to leave the ordering unchanged.}

\item{nbrClustersCols}{number of clusters in columns separated by gaps.}

\item{colorAnnoRow}{row annotations shown left from the heatmap, as a character vector. If \code{NA} (default), all annotations are shown.}

\item{colorAnnoCol}{column annotations shown above the heatmap, as a character vector. If \code{NA} (default), all annotations are shown.}

\item{legendColorScheme}{coloring scheme used for the annotations. One from \code{"Accent"}, \code{"Dark2"}, \code{"Paired"}, \code{"Pastel1"}, \code{"Pastel2"}, \code{"Set1"} (default), \code{"Set2"} or \code{"Set3"}.}

\item{plotWidth}{width of the plot.}

\item{plotRatio}{ratio of height and width of the image.}

\item{colorRangeMin}{minimum of the color range or \code{NA} (default) to calculate it automatically.}

\item{colorRangeMax}{maximum of the color range or \code{NA} (default) to calculate it automatically.}

\item{matrixColorScheme}{coloring scheme used for the heatmap. One of the sequential or diverging color schemes from \code{"RColocBrewer"} package, see row names of \code{"RColorBrewer::brewer.pal.info"} for the list.}

\item{revScheme}{whether to reverse the coloring scheme.}

\item{cellBorder}{color of the cell border or \code{NA} to omit border.}

\item{fontSizeGeneral}{general font size.}

\item{showNumbers}{whether to show numbers in the cells.}

\item{fontSizeNumbers}{font size of the numbers.}

\item{precisionNumbers}{number of decimal places shown.}

\item{showRownames}{whether to show row names.}

\item{fontSizeRownames}{font size of the row names.}

\item{showColnames}{whether to show column names.}

\item{fontSizeColnames}{font size of the column names.}

\item{showAnnoTitlesRow}{whether to show row annotation titles.}

\item{showAnnoTitlesCol}{whether to show column annotation titles.}

\item{maxAnnoLevels}{maximum number of annotation levels.}
}
\value{
a structure to be used as input for the function \code{saveHeatmap}.
}
\description{
This function performs the steps shown on the 'Heatmap' tab of the online ClustVis.
}
