% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{generatePCA}
\alias{generatePCA}
\title{Generate ClustVis PCA plot.}
\usage{
generatePCA(proc, pcx = 1, pcy = 2, switchDirX = FALSE,
  switchDirY = FALSE, colorAnno = 1, colorScheme = "Set1",
  showEllipses = TRUE, ellipseConf = 0.95, ellipseLineWidth = 1,
  ellipseLineType = "solid", shapeAnno = 2, shapeScheme = "various",
  plotWidth = 20, plotRatio = 0.8, marginRatio = 0.05, pointSize = 5,
  legendPosition = "right", fontSize = 20, axisLabelPrefix = "PC",
  showVariance = TRUE, showSampleIds = FALSE, maxColorLevels = 8,
  maxShapeLevels = 62)
}
\arguments{
\item{proc}{structure returned by the \code{processData} function.}

\item{pcx}{principal component shown on the x-axis.}

\item{pcy}{principal component shown on the y-axis.}

\item{switchDirX}{whether to switch the direction of x-axis (reverse the sign).}

\item{switchDirY}{whether to switch the direction of y-axis (reverse the sign).}

\item{colorAnno}{the annotation(s) used for coloring and ellipses, as column numbers (if numeric) or column names (if character). If length is greater than one, unique combinations of the levels are found. If \code{NULL}, all points are black.}

\item{colorScheme}{coloring scheme used for the annotation groups. One from \code{"Accent"}, \code{"Dark2"}, \code{"Paired"}, \code{"Pastel1"}, \code{"Pastel2"}, \code{"Set1"} (default), \code{"Set2"}, \code{"Set3"} or \code{"Grayscale"}.}

\item{showEllipses}{whether to show prediction ellipses around groups.}

\item{ellipseConf}{confidence level used for calculating prediction ellipses.}

\item{ellipseLineWidth}{line width for the ellipses.}

\item{ellipseLineType}{line type for the ellipses. One from \code{"solid"} (default), \code{"dashed"}, \code{"dotted"}, \code{"dotdash"}, \code{"longdash"} or \code{"twodash"}.}

\item{shapeAnno}{the annotation(s) used for shape, as column numbers (if numeric) or column names (if character). If length is greater than one, unique combinations of the levels are found. If \code{NULL}, all points are round.}

\item{shapeScheme}{shape scheme used for annotation groups. One from \code{"various"} (default), \code{"letters"}, or a list of point symbol codes or characters.}

\item{plotWidth}{plot width in cm.}

\item{plotRatio}{ratio of height and width of the plotting area.}

\item{marginRatio}{how much white space is shown between data and plot borders, relative to y-axis data range. Margin is added after forcing the specified plot ratio.}

\item{pointSize}{point size.}

\item{legendPosition}{legend position relative to the plot.}

\item{fontSize}{font size.}

\item{axisLabelPrefix}{prefix shown for the axis labels.}

\item{showVariance}{whether to show percentage of variance explained by the component in the axis label.}

\item{showSampleIds}{whether to show sample IDs for the points shown on the plot.}

\item{maxColorLevels}{maximum number of levels allowed for colors.}

\item{maxShapeLevels}{maximum number of levels allowed for shapes.}
}
\value{
a structure to be used as input for the function \code{savePCA}.
}
\description{
This function performs the steps shown on the 'PCA' tab of the online ClustVis.
}
