% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{importData}
\alias{importData}
\title{Import data into ClustVis.}
\usage{
importData(file, sep = NA, nbrRowAnnos = NA, nbrColAnnos = NA,
  quotes = "\\"'", naString = "NA", filteringRows = NULL,
  filteringCols = NULL, transpose = FALSE)
}
\arguments{
\item{file}{file name containing the input data.}

\item{sep}{field separator between columns. If \code{NA} (default), the separator is detected automatically.}

\item{nbrRowAnnos}{number of row annotations placed as the first columns in the input file. If \code{nbrRowAnnos} or \code{nbrColAnnos} is \code{NA} (default), number of row annotations is detected automatically.}

\item{nbrColAnnos}{number of column annotations placed as the first rows in the input file. If \code{nbrRowAnnos} or \code{nbrColAnnos} is \code{NA} (default), number of column annotations is detected automatically.}

\item{quotes}{quoting characters. See \code{quote} parameter from \code{read.table} function for more details.}

\item{naString}{string for representing a missing value. See \code{na.srings} parameter from \code{read.table} function for more details.}

\item{filteringRows}{logical vector showing which rows to keep. Should have the same length as the number of rows in the data matrix. \code{NULL} (default) keeps all rows.}

\item{filteringCols}{logical vector showing which columns to keep. Should have the same length as the number of columns in the data matrix. \code{NULL} (default) keeps all columns.}

\item{transpose}{whether to transpose the data matrix.}
}
\value{
a structure to be used as input for the function \code{processData}
}
\description{
This function performs the steps shown on the 'Data import' tab of the online ClustVis.
}
