% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{processData}
\alias{processData}
\title{Pre-process ClustVis data.}
\usage{
processData(data, transformation = NA, annoColKeep = NULL,
  annoColMethodAgg = NA, maxNaRows = 0.9999, maxNaCols = 0.9999,
  remConstCols = FALSE, rowCentering = TRUE, rowScaling = "uv",
  pcaMethod = "svdImpute", maxComponents = 100)
}
\arguments{
\item{data}{data structure returned by the \code{importData} function.}

\item{transformation}{\code{"ln(x)"} or \code{"ln(x + 1)"} to apply logarithmic tranformation, or \code{NA} (default) to use original values.}

\item{annoColKeep}{column names to keep or \code{NULL} (default) to keep all columns.}

\item{annoColMethodAgg}{column aggregation method (function name like \code{"mean"} or \code{"median"} as a string) or \code{NA} (default) to keep original columns without aggregation.}

\item{maxNaRows}{maximum proportion of missing values allowed in rows. Rows having a higher proportion of missing values are removed from further analysis.}

\item{maxNaCols}{maximum proportion of missing values allowed in columns. Columns having a higher proportion of missing values are removed from further analysis.}

\item{remConstCols}{whether to remove constant columns.}

\item{rowCentering}{whether to center rows to have zero mean.}

\item{rowScaling}{which method to use for scaling the rows. One from \code{"uv"} (default), \code{"pareto"}, \code{"vector"} or \code{"none"} for keeping the rows without scaling.}

\item{pcaMethod}{which method to use for calculating principal components. One from \code{"svdImpute"} (default), \code{"nipals"} or \code{"ppca"}.}

\item{maxComponents}{maximum number of PCA components to calculate.}
}
\value{
a structure to be used as input for the functions \code{generatePCA} and \code{generateHeatmap}.
}
\description{
This function performs the steps shown on the 'Data pre-processing' tab of the online ClustVis.
}
