% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{saveHeatmap}
\alias{saveHeatmap}
\title{Save ClustVis heatmap plot.}
\usage{
saveHeatmap(hm, file = NA)
}
\arguments{
\item{hm}{structure returned by the \code{generateHeatmap} function.}

\item{file}{relative or absolute path of the output file, or \code{NA} (default) for current output device.}
}
\description{
This function saves the heatmap output.
}
