% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustvis.R
\name{savePCA}
\alias{savePCA}
\title{Save ClustVis PCA plot.}
\usage{
savePCA(pca, file = NA)
}
\arguments{
\item{pca}{structure returned by the \code{generatePCA} function.}

\item{file}{relative or absolute path of the output file, or \code{NA} (default) for current output device.}
}
\description{
This function saves the PCA output.
}
